adjusting table structure to match others
